style: Historian 클래스 Histories 로 이름 변경
