+++ README.md
@@ -46,6 +46,45 @@
 <hr/>
 ## ✅ 피드백 반영사항  <!-- 지난 코드리뷰에서 고친 사항을 적어주세요. 재PR 시에만 사용해 주세요! (재PR 아닌 경우 삭제) -->
 
+`계산에 들어갈 요소가 어떤게 있을지 생각하면서 적은게 좋았다.`
+#### 구조
+* io 패키지 제외 모두 model 패키지에 들어갈 수 있음.
+* `Historian` -> `Histories`
+* Input, Output도 불필요한 추상화 들어감.
+* Computer, Converter를 interface로 분리한 이유가 없음.
+
+#### 일반
+* 컬렉션은 복수 표기가 일반적.
+* `exp`같은 축약어 사용 지양하기
+* 자바 컨벤션에서는 중괄호 생략하지 않음.
+
+#### 구현
+* `Calculator` 클래스에서 `label:` 구문 제거
+    * 향상된 `switch` 문으로 `break` 키워드 제거
+    * `case` 조건 String을 다시 `Menu enum` 클래스로 변환
+* `Calculator` `run` 메소드에서 들여쓰기 조절
+    * run에는 while 문만 두기
+    * actual operate에 `try-catch` 블럭 두기
+    * menu operate에서 향상 switch 문으로 유저 상호작용하기
+* `validator`가 모호
+    * 검증하는 객체인데 검증, 변환 둘 다 함
+    * 검증만 수행하도록 분리
+* `SimpleValidator` 클래스에 있는 필드 변수 책임
+    * 생성자로 초기화해주던가
+    * 초기화하는 기능을 validator 바깥에 따로 해주던가
+* `SimpleComputer` `calculate` 메소드 사용부분 단순화
+    * 변수로 빼는 방법
+    * 메소드로 나누는 방법
+* `Historian` 클래스에서 띄어쓰기, 수식 넣는걸 `output`으로 책임 넘기기
+* 괄호는 operator인가?
+* `Operator` 클래스 `getOperator` 메소드는 값을 못 찾을 때 NPE 발생 가능
+    * get(0) 말고도, orelse로 예외처리?
+    * Optional로 받을까?
+
+#### 테스트
+* 테스트 코드 given, when, then 형태로 나누기
+* @ParameterizedTest 이용 -> 테스트 코드 중복 줄이기
+
 <hr/>
 ## ✅ PR 포인트 & 궁금한 점 <!-- 리뷰어 분들이 집중적으로 보셨으면 하는 내용을 적어주세요 -->
 1. 테스트 코드
