+++ src/main/java/com/devcourse/App.java
@@ -1,12 +1,12 @@
 package com.devcourse;
 
-import com.devcourse.engine.computer.SimpleComputer;
+import com.devcourse.engine.model.computer.SimpleComputer;
 import com.devcourse.engine.Calculator;
-import com.devcourse.engine.converter.PostfixConverter;
-import com.devcourse.engine.historian.Historian;
+import com.devcourse.engine.model.converter.PostfixConverter;
+import com.devcourse.engine.model.histories.Histories;
 import com.devcourse.engine.io.InputConsole;
 import com.devcourse.engine.io.OutputConsole;
-import com.devcourse.engine.validator.SimpleValidator;
+import com.devcourse.engine.model.validator.SimpleValidator;
 
 public class App {
     public static void main(String[] args) {
@@ -14,7 +14,7 @@ public class App {
         new Calculator(
                 new InputConsole(),
                 new OutputConsole(),
-                new Historian(),
+                new Histories(),
                 new SimpleValidator(),
                 new PostfixConverter(),
                 new SimpleComputer()
